# method_pipeline/generated/chatdev/minimax-m2.7/generic/QA2-1207/sdk_chatdev_20260712091919_20260712091919/code_workspace/oms_backend/openapi.yaml
openapi: 3.0.3
info:
  title: Order Management System API
  description: |
    A production-grade Order Management System backend implementing the complete order workflow.
    
    ## Workflow Steps
    1. Customer places order
    2. Order Staff reviews & accepts
    3. Accountant creates invoice
    4. Customer pays invoice
    5. Accountant verifies payment
    6. Order Staff ships paid order
    7. Order Staff closes completed order
    
    ## NFR Support
    - NFR 2.1 Graceful Degradation: Feature flags and circuit breakers
    - NFR 2.2 Fault Detection: Health checks and automatic recovery
    - NFR 2.3 State Preservation: WAL mode, snapshots, and idempotency keys
  version: 1.0.0
  contact:
    name: OMS Backend Team
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/v1/products:
    get:
      tags: [products]
      summary: List products
      parameters:
        - name: skip
          in: query
          schema: {type: integer, default: 0}
        - name: limit
          in: query
          schema: {type: integer, default: 100}
        - name: active_only
          in: query
          schema: {type: boolean, default: false}
      responses:
        '200':
          description: List of products
    post:
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created
  /api/v1/products/{product_id}:
    get:
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: product_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Product found
  /api/v1/orders:
    get:
      tags: [orders]
      summary: List orders
      parameters:
        - name: skip
          in: query
          schema: {type: integer, default: 0}
        - name: limit
          in: query
          schema: {type: integer, default: 100}
        - name: status
          in: query
          schema: {type: string}
        - name: customer_id
          in: query
          schema: {type: string}
      responses:
        '200':
          description: List of orders
    post:
      tags: [orders]
      summary: Place order (Step 1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
      responses:
        '201':
          description: Order placed
  /api/v1/orders/{order_id}:
    get:
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: order_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Order found
  /api/v1/orders/{order_id}/accept:
    patch:
      tags: [orders]
      summary: Accept order (Step 2)
      parameters:
        - name: order_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Order accepted
  /api/v1/orders/{order_id}/ship:
    patch:
      tags: [orders]
      summary: Ship order (Step 6)
      parameters:
        - name: order_id
          in: path
          required: true
          schema: {type: string}
        - name: tracking_number
          in: query
          schema: {type: string}
      responses:
        '200':
          description: Order shipped
  /api/v1/orders/{order_id}/close:
    patch:
      tags: [orders]
      summary: Close order (Step 7)
      parameters:
        - name: order_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Order closed
...
  /api/v1/payments:
    post:
      tags: [payments]
      summary: Create payment (Step 4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment created
  /api/v1/payments/{payment_id}/process:
    post:
      tags: [payments]
      summary: Process payment
      parameters:
        - name: payment_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Payment processed
  /api/v1/payments/{payment_id}/verify:
    post:
      tags: [payments]
      summary: Verify payment (Step 5)
      parameters:
        - name: payment_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Payment verified
...
components:
  schemas:
...
    Address:
      type: object
      properties:
        street: {type: string}
        city: {type: string}
        state: {type: string}
        postal_code: {type: string}
        country: {type: string}
...
    CreateProductRequest:
      type: object
      required: [sku, base_price]
      properties:
        sku: {type: string}
        description: {type: string}
        base_price: {type: number}
        currency: {type: string, default: USD}
        stock_quantity: {type: integer, default: 0}
...
    PlaceOrderRequest:
      type: object
      required: [customer_id, line_items, shipping_address]
      properties:
        customer_id: {type: string}
        line_items: {type: array}
        shipping_address: {$ref: '#/components/schemas/Address'}
        notes: {type: string}
        idempotency_key: {type: string}
...
    CreatePaymentRequest:
      type: object
      required: [order_id, invoice_id, customer_id, amount]
      properties:
        order_id: {type: string}
        invoice_id: {type: string}
        customer_id: {type: string}
        amount: {type: number}
        method: {type: string, default: bank_transfer}
        idempotency_key: {type: string}